requirements.txt
django==5.0.4
psycopg2-binary==2.9.9
django-environ==0.11.2
djangorestframework==3.15.1
django-import-export==3.3.8
openpyxl==3.1.2
